refactor/tweaks
